Made game demo more readable
